Module Owi

module Abstract_boolean : sig ... end

module Abstract_checker : sig ... end

module Abstract_data : sig ... end

module Abstract_domain : sig ... end

module Abstract_elem : sig ... end

module Abstract_extern : sig ... end

module Abstract_f32 : sig ... end

module Abstract_f64 : sig ... end

module Abstract_i32 : sig ... end

module Abstract_i64 : sig ... end

module Abstract_interpreter_control_flow : sig ... end

module Abstract_interpreter_simple : sig ... end

module Abstract_interpreter_state : sig ... end

module Abstract_invariant : sig ... end

module Abstract_jump_map : sig ... end

module Abstract_locals : sig ... end

module Abstract_memory : sig ... end

module Abstract_monad : sig ... end

module Abstract_ref : sig ... end

module Abstract_stack : sig ... end

module Abstract_state : sig ... end

module Abstract_table : sig ... end

module Abstract_trace : sig ... end

module Abstract_v128 : sig ... end

module Abstract_value : sig ... end

module Abstract_wasm_ffi : sig ... end

module Allocator : sig ... end

module Annotated : sig ... end

module Array_intf : sig ... end

module Assigned : sig ... end

module Benchmark : sig ... end

module Binary : sig ... end

module Binary_encoder : sig ... end

module Binary_parser : sig ... end

module Binary_to_text : sig ... end

module Binary_validate : sig ... end

Module to typecheck a simplified module.

module Boolean_intf : sig ... end

module Bug : sig ... end

module Call_graph : sig ... end

module Choice_intf : sig ... end

module Cmd_c : sig ... end

module Cmd_cpp : sig ... end

module Cmd_go : sig ... end

module Cmd_haskell : sig ... end

module Cmd_llvm : sig ... end

module Cmd_rust : sig ... end

module Cmd_utils : sig ... end

module Cmd_version : sig ... end

module Cmd_wasm_abs : sig ... end

module Cmd_wasm_fmt : sig ... end

module Cmd_wasm_fuzz : sig ... end

module Cmd_wasm_hunt : sig ... end

module Cmd_wasm_inspect_cfg : sig ... end

module Cmd_wasm_inspect_cg : sig ... end

module Cmd_wasm_instrument_label : sig ... end

module Cmd_wasm_iso : sig ... end

module Cmd_wasm_of_wat : sig ... end

module Cmd_wasm_replay : sig ... end

module Cmd_wasm_run : sig ... end

module Cmd_wasm_script : sig ... end

module Cmd_wasm_sym : sig ... end

module Cmd_wasm_to_wat : sig ... end

module Cmd_wasm_validate : sig ... end

module Cmd_zig : sig ... end

module Collection : sig ... end

module Compile : sig ... end

module Concrete_array : sig ... end

module Concrete_boolean : sig ... end

module Concrete_choice : sig ... end

module Concrete_data : sig ... end

module Concrete_elem : sig ... end

module Concrete_extern : sig ... end

module Concrete_f32 : sig ... end

module Concrete_f64 : sig ... end

module Concrete_i16 : sig ... end

module Concrete_i32 : sig ... end

module Concrete_i64 : sig ... end

module Concrete_i8 : sig ... end

module Concrete_memory : sig ... end

module Concrete_ref : sig ... end

module Concrete_state : sig ... end

module Concrete_struct : sig ... end

module Concrete_table : sig ... end

runtime table

module Concrete_v128 : sig ... end

module Concrete_value : sig ... end

module Concrete_wasm_ffi : sig ... end

module Constexpr_eval : sig ... end

module Constexpr_eval_intf : sig ... end

module Control_flow_graph : sig ... end

module Data_intf : sig ... end

module Elem_intf : sig ... end

module Env : sig ... end

module Env0 : sig ... end

module Env_intf : sig ... end

module Env_linker : sig ... end

module Env_rewriter : sig ... end

module Exploration_smart : sig ... end

module Extern : sig ... end

module Extern_intf : sig ... end

module F32_intf : sig ... end

module F64_intf : sig ... end

module Float32 : sig ... end

Custom Float32 module for Wasm.

module Float64 : sig ... end

Custom Float64 module for Wasm.

module Fop_intf : sig ... end

module Fuzz_dict : sig ... end

module Fuzz_driver : sig ... end

module Fuzz_gen : sig ... end

module Fuzz_wasm_ffi : sig ... end

module Grouped : sig ... end

module I16_intf : sig ... end

module I32_intf : sig ... end

module I64_intf : sig ... end

module I8_intf : sig ... end

module Init : sig ... end

module Int32 : sig ... end

Custom Int32 module for Wasm.

module Int64 : sig ... end

Custom Int64 module for Wasm.

module Interpret : sig ... end

module Iop_intf : sig ... end

module Kind : sig ... end

module Label : sig ... end

module Log : sig ... end

module Memory_intf : sig ... end

module Model : sig ... end

module Multicore : sig ... end

module Origin : sig ... end

module Parse : sig ... end

Module providing functions to parse a wasm script from various kind of inputs.

module Prio : sig ... end

module Priority_queue : sig ... end

module Ref_intf : sig ... end

module Result : sig ... end

module Rewrite : sig ... end

module Script : sig ... end

Module to execute a full Wasm script.

module Script_abstract : sig ... end

module Script_error : sig ... end

module Script_symbolic : sig ... end

Module to execute a full Wasm script.

module Share : sig ... end

module Solver : sig ... end

module Spectest : sig ... end

The `spectest` module, to run script from the official test suite.

module Stack : sig ... end

module Struct_intf : sig ... end

module Symbol_scope : sig ... end

module Symbolic_array : sig ... end

module Symbolic_boolean : sig ... end

module Symbolic_choice : sig ... end

module Symbolic_data : sig ... end

module Symbolic_driver : sig ... end

module Symbolic_elem : sig ... end

module Symbolic_extern : sig ... end

module Symbolic_f32 : sig ... end

module Symbolic_f64 : sig ... end

module Symbolic_i16 : sig ... end

module Symbolic_i32 : sig ... end

module Symbolic_i64 : sig ... end

module Symbolic_i8 : sig ... end

module Symbolic_memory : sig ... end

module Symbolic_parameters : sig ... end

module Symbolic_ref : sig ... end

module Symbolic_struct : sig ... end

module Symbolic_table : sig ... end

module Symbolic_v128 : sig ... end

module Symbolic_value : sig ... end

module Symbolic_wasm_ffi : sig ... end

module Syntax : sig ... end

module Table_intf : sig ... end

module Text : sig ... end

module Text_keywords : sig ... end

module Text_lexer : sig ... end

Module for Wasm lexing.

module Text_parser : sig ... end

module Text_validate : sig ... end

Initial check done on a module.

module Thread : sig ... end

module V128_intf : sig ... end

module Value_intf : sig ... end

module Wasm_ffi_intf : sig ... end

module Wast : sig ... end

module Wutf8 : sig ... end

Utility functions to work with utf8.