Module Binary.Module

type t = {

id : string option;

types : Types.binary Types.type_def array;

global : (global, Types.binary Types.global_type) Runtime.t array;

table : (Types.binary Types.table, Types.binary Types.table_type) Runtime.t array;

mem : (Types.mem, Types.limits) Runtime.t array;

func : (Types.binary Types.func, Types.binary Types.block_type) Runtime.t array;

elem : elem array;

data : data array;

exports : exports;

start : int option;

custom : custom list;

}

val empty : t

Exports

val find_exported_func_from_name : string -> t -> export option

Return the first function exported as name if it exists. Return None otherwise.

Imports

val find_imported_func_index : modul_name:string -> func_name:string -> t -> int option

Return the index of a function imported from a given modul_name and func_name if it exists. Return None otherwise.

val add_import_if_not_present : modul_name:string -> func_name:string -> desc:Types.binary Types.block_type -> t -> < .. > Types.indice * t

Look for an imported function index, adding it if not already imported.

Functions

val add_func : (Types.binary Types.func, Types.binary Types.block_type) Runtime.t -> t -> t * int

Add a function f to a module m and returns the module and the index of the added function.

val get_func_type : int -> t -> Types.binary Types.block_type option

Return the type of the function at index id.