Module Compile.Binary

val until_binary_validate : unsafe:bool -> Binary.Module.t -> Binary.Module.t Result.t

val until_optimize : unsafe:bool -> optimize:bool -> Binary.Module.t -> Binary.Module.t Result.t

val until_link : unsafe:bool -> optimize:bool -> name:string option -> 'f Link.state -> Binary.Module.t -> ('f Link.module_to_run * 'f Link.state) Result.t

compile a module with a given link state and produce a new link state and a runnable module

val until_interpret : unsafe:bool -> optimize:bool -> name:string option -> Concrete_extern_func.extern_func Link.state -> Binary.Module.t -> Concrete_extern_func.extern_func Link.state Result.t

compile and interpret a module with a given link state and produce a new link state