Module Compile.File

val until_validate : unsafe:bool -> rac:bool -> srac:bool -> Fpath.t -> Binary.Module.t Result.t

val until_link : unsafe:bool -> rac:bool -> srac:bool -> name:string option -> 'extern_func Link.state -> Fpath.t -> ('extern_func Link.module_to_run * 'extern_func Link.state) Result.t

compile a file with a given link state and produce a new link state and a runnable module