Module Wasm.File

val until_binary : unsafe:bool -> Fpath.t -> Binary.Module.t Result.t

val until_validate : unsafe:bool -> Fpath.t -> Binary.Module.t Result.t

val until_concrete_link : unsafe:bool -> name:string option -> Env.Concrete.t -> Fpath.t -> (Env.Concrete.modul * Env.Concrete.t) Result.t

compile a file with a given link state and produce a new link state and a runnable module

val until_symbolic_link : unsafe:bool -> name:string option -> Env.Symbolic.t -> Fpath.t -> (Env.Symbolic.modul * Env.Symbolic.t) Result.t

compile a file with a given link state and produce a new link state and a runnable module

val until_abstract_link : unsafe:bool -> name:string option -> Env.Abstract.t -> Fpath.t -> (Env.Abstract.modul * Env.Abstract.t) Result.t

compile a file with a given link state and produce a new link state and a runnable module