Module Owi.Link

Module to link a simplified/extern module and producing a runnable module along with a link state.

module Memory : sig ... end

runtime memory

module Table : sig ... end

runtime table

module Global : sig ... end

runtime global

module Env : sig ... end

runtime env

type module_to_run = {

module_ : Simplify.simplified_module;

env : Env.t;

to_run : (int, Types.func_type) Types.expr' list;

}

runnable module

module StringMap : Stdlib.Map.S with type key = string

module StringSet : Stdlib.Set.S

type exports = {

globals : Env.t' Global.t StringMap.t;

memories : Memory.t StringMap.t;

tables : Env.t' Table.t StringMap.t;

functions : Env.t' Value.Func.t StringMap.t;

defined_names : StringSet.t;

}

runtime exported items

type state = {

by_name : exports StringMap.t;

by_id : exports StringMap.t;

last : exports option;

}

link state

val empty_state : state

the empty link state

val module_ : Simplify.simplified_module -> state -> (module_to_run * state, string) Stdlib.result

link a module with a given link state, producing a runnable module and a new link state

val register_module : state -> name:string -> id:string option -> (state, string) Stdlib.Result.t

register a module inside a link state, producing a new link state

type extern_module = {

functions : (string * Value.Func.extern_func) list;

}

extern modules

val extern_module : string -> extern_module -> state -> state

register an extern module with a given link state, producing a new link state