Module Owi.Spectest

The `spectest` module, to run script from the official test suite.

val extern_m : Concrete_extern.Module.t

val symbolic_extern_m : Symbolic_extern.Module.t

val abstract_extern_m : Abstract_extern.Module.t

val m : Wast.cmd

the spectest module, which is simply a Wast Text module importing everything required