Parameter Make._

type t

type address

val address : Smtml.Expr.t -> address Symbolic_choice_without_memory.t

val address_i32 : Int32.t -> address

val make : unit -> t

val clone : t -> t

val loadn : t -> address -> int -> Smtml.Expr.t

val storen : t -> address -> Smtml.Expr.t -> int -> unit

val validate_address : t -> Smtml.Expr.t -> int -> Smtml.Expr.t Result.t Symbolic_choice_without_memory.t

validate_address m a range verifies whether an operation starting at address a is valid within the address range a to a + range - 1 (inclusive).

val realloc : t -> ptr:Smtml.Expr.t -> size:Smtml.Expr.t -> Smtml.Expr.t Symbolic_choice_without_memory.t

val free : t -> Smtml.Expr.t -> Smtml.Expr.t Symbolic_choice_without_memory.t