Module Owi.Text

exception Parse_fail of string

val sp : Stdlib.Format.formatter -> unit -> unit

type indice =

| Text of string

| Raw of int

val pp_id : Stdlib.Format.formatter -> string -> unit

val pp_id_opt : Stdlib.Format.formatter -> string option -> unit

val pp_indice : Stdlib.Format.formatter -> indice -> unit

val compare_indice : indice -> indice -> int

val indice_eq : indice -> indice -> bool

val pp_indice_opt : Stdlib.Format.formatter -> indice option -> unit

val pp_indices : Stdlib.Format.formatter -> indice list -> unit

type nonrec num_type =

| I32

| I64

| F32

| F64

| V128

val pp_num_type : Stdlib.Format.formatter -> num_type -> unit

val num_type_eq : num_type -> num_type -> bool

val compare_num_type : num_type -> num_type -> int

type nullable =

| No_null

| Null

val pp_nullable : Stdlib.Format.formatter -> nullable -> unit

type nonrec mut =

| Const

| Var

val pp_mut : Stdlib.Format.formatter -> mut -> unit

val is_mut : mut -> bool

type nonrec nn =

| S32

| S64

val pp_nn : Stdlib.Format.formatter -> nn -> unit

type nonrec ishape =

| I8x16

| I16x8

| I32x4

| I64x2

val pp_ishape : Stdlib.Format.formatter -> ishape -> unit

type nonrec fshape =

| F32x4

| F64x8

val pp_fshape : Stdlib.Format.formatter -> fshape -> unit

type nonrec sx =

| U

| S

val pp_sx : Stdlib.Format.formatter -> sx -> unit

type nonrec iunop =

| Clz

| Ctz

| Popcnt

val pp_iunop : Stdlib.Format.formatter -> iunop -> unit

type nonrec funop =

| Abs

| Neg

| Sqrt

| Ceil

| Floor

| Trunc

| Nearest

val pp_funop : Stdlib.Format.formatter -> funop -> unit

type nonrec vibinop =

| Add

| Sub

val pp_vibinop : Stdlib.Format.formatter -> vibinop -> unit

type nonrec ibinop =

| Add

| Sub

| Mul

| Div of sx

| Rem of sx

| And

| Or

| Xor

| Shl

| Shr of sx

| Rotl

| Rotr

val pp_ibinop : Stdlib.Format.formatter -> ibinop -> unit

type nonrec fbinop =

| Add

| Sub

| Mul

| Div

| Min

| Max

| Copysign

val pp_fbinop : Stdlib.Format.formatter -> fbinop -> unit

type nonrec itestop =

| Eqz

val pp_itestop : Stdlib.Format.formatter -> itestop -> unit

type nonrec irelop =

| Eq

| Ne

| Lt of sx

| Gt of sx

| Le of sx

| Ge of sx

val pp_irelop : Stdlib.Format.formatter -> irelop -> Prelude.Unit.t

type nonrec frelop =

| Eq

| Ne

| Lt

| Gt

| Le

| Ge

val frelop : Stdlib.Format.formatter -> frelop -> Prelude.Unit.t

type nonrec memarg = {

offset : Int32.t;

align : Int32.t;

}

val pp_memarg : Stdlib.Format.formatter -> memarg -> unit

type nonrec limits = {

min : int;

max : int option;

}

val pp_limits : Stdlib.Format.formatter -> limits -> unit

type nonrec mem = string option * limits

val pp_mem : Stdlib.Format.formatter -> (string option * limits) -> unit

Structure

Types

type heap_type =

| Func_ht

| Extern_ht

val pp_heap_type : Stdlib.Format.formatter -> heap_type -> unit

val pp_heap_type_short : Stdlib.Format.formatter -> heap_type -> unit

val heap_type_eq : heap_type -> heap_type -> bool

val compare_heap_type : heap_type -> heap_type -> int

type nonrec ref_type = nullable * heap_type

val pp_ref_type : Stdlib.Format.formatter -> (nullable * heap_type) -> unit

val ref_type_eq : (nullable * heap_type) -> (nullable * heap_type) -> bool

val compare_ref_type : (nullable * heap_type) -> (nullable * heap_type) -> int

type nonrec val_type =

| Num_type of num_type

| Ref_type of ref_type

val pp_val_type : Stdlib.Format.formatter -> val_type -> unit

val val_type_eq : val_type -> val_type -> bool

val compare_val_type : val_type -> val_type -> int

type nonrec param = string option * val_type

val pp_param : Stdlib.Format.formatter -> (string option * val_type) -> unit

val param_eq : ('a * val_type) -> ('b * val_type) -> bool

val compare_param : ('a * val_type) -> ('b * val_type) -> int

type nonrec param_type = param list

val pp_param_type : Stdlib.Format.formatter -> (string option * val_type) list -> unit

val param_type_eq : ('a * val_type) list -> ('a * val_type) list -> bool

val compare_param_type : ('a * val_type) list -> ('a * val_type) list -> int

type nonrec result_type = val_type list

val pp_result_ : Stdlib.Format.formatter -> val_type -> unit

val pp_result_type : Stdlib.Format.formatter -> val_type list -> unit

val result_type_eq : val_type list -> val_type list -> bool

val compare_result_type : val_type list -> val_type list -> int

val with_space_list : (Stdlib.Format.formatter -> 'a list -> unit) -> Stdlib.Format.formatter -> 'a list -> unit

type nonrec func_type = param_type * result_type

val pp_func_type : Stdlib.Format.formatter -> ((string option * val_type) list * val_type list) -> unit

val func_type_eq : (('a * val_type) list * val_type list) -> (('a * val_type) list * val_type list) -> bool

type block_type =

| Bt_ind of indice

| Bt_raw of indice option * func_type

val pp_block_type : Stdlib.Format.formatter -> block_type -> unit

val pp_block_type_opt : Stdlib.Format.formatter -> block_type option -> unit

val compare_func_type : (('a * val_type) list * val_type list) -> (('a * val_type) list * val_type list) -> int

type nonrec table_type = limits * ref_type

val pp_table_type : Stdlib.Format.formatter -> (limits * (nullable * heap_type)) -> unit

type nonrec global_type = mut * val_type

val pp_global_type : Stdlib.Format.formatter -> (mut * val_type) -> unit

type nonrec extern_type =

| Func of string option * func_type

| Table of string option * table_type

| Mem of string option * limits

| Global of string option * global_type

Instructions

type instr =

| I32_const of Int32.t

| I64_const of Int64.t

| F32_const of Float32.t

| F64_const of Float64.t

| V128_const of V128.t

| I_unop of nn * iunop

| F_unop of nn * funop

| I_binop of nn * ibinop

| F_binop of nn * fbinop

| V_ibinop of ishape * vibinop

| I_testop of nn * itestop

| I_relop of nn * irelop

| F_relop of nn * frelop

| I_extend8_s of nn

| I_extend16_s of nn

| I64_extend32_s

| I32_wrap_i64

| I64_extend_i32 of sx

| I_trunc_f of nn * nn * sx

| I_trunc_sat_f of nn * nn * sx

| F32_demote_f64

| F64_promote_f32

| F_convert_i of nn * nn * sx

| I_reinterpret_f of nn * nn

| F_reinterpret_i of nn * nn

| Ref_null of heap_type

| Ref_is_null

| Ref_func of indice

| Drop

| Select of val_type list option

| Local_get of indice

| Local_set of indice

| Local_tee of indice

| Global_get of indice

| Global_set of indice

| Table_get of indice

| Table_set of indice

| Table_size of indice

| Table_grow of indice

| Table_fill of indice

| Table_copy of indice * indice

| Table_init of indice * indice

| Elem_drop of indice

| I_load of nn * memarg

| F_load of nn * memarg

| I_store of nn * memarg

| F_store of nn * memarg

| I_load8 of nn * sx * memarg

| I_load16 of nn * sx * memarg

| I64_load32 of sx * memarg

| I_store8 of nn * memarg

| I_store16 of nn * memarg

| I64_store32 of memarg

| Memory_size

| Memory_grow

| Memory_fill

| Memory_copy

| Memory_init of indice

| Data_drop of indice

| Nop

| Unreachable

| Block of string option * block_type option * expr Annotated.t

| Loop of string option * block_type option * expr Annotated.t

| If_else of string option * block_type option * expr Annotated.t * expr Annotated.t

| Br of indice

| Br_if of indice

| Br_table of indice array * indice

| Return

| Return_call of indice

| Return_call_indirect of indice * block_type

| Return_call_ref of block_type

| Call of indice

| Call_indirect of indice * block_type

| Call_ref of indice

| Extern_externalize

| Extern_internalize

and expr = instr Annotated.t list

val pp_newline : Stdlib.Format.formatter -> unit -> unit

val pp_instr : short:bool -> Stdlib.Format.formatter -> instr -> Prelude.Unit.t

val pp_expr : short:bool -> Stdlib.Format.formatter -> expr Annotated.t -> Prelude.Unit.t

val iter_expr : (instr -> Prelude.Unit.t) -> expr Annotated.t -> Prelude.Unit.t

val iter_instr : (instr -> Prelude.Unit.t) -> instr Annotated.t -> Prelude.Unit.t

type func = {

type_f : block_type;

locals : param list;

body : expr Annotated.t;

id : string option;

}

val pp_local : Stdlib.Format.formatter -> (string option * val_type) -> unit

val pp_locals : Stdlib.Format.formatter -> (string option * val_type) list -> unit

val pp_func : Stdlib.Format.formatter -> func -> Prelude.Unit.t

val pp_funcs : Stdlib.Format.formatter -> func list -> unit

type table = string option * table_type

val pp_table : Stdlib.Format.formatter -> (string option * (limits * (nullable * heap_type))) -> unit

type import_desc =

| Import_func of string option * block_type

| Import_table of string option * table_type

| Import_mem of string option * limits

| Import_global of string option * global_type

val import_desc : Stdlib.Format.formatter -> import_desc -> unit

type import = {

modul : string;
(*
The name of the module from which the import is done
*)

name : string;
(*
The name of the importee in its module of origin
*)

desc : import_desc;
(*
If this import_desc first field is Some s, the importee is made available under name s, else it can only be used via its numerical index.
*)

}

val pp_import : Stdlib.Format.formatter -> import -> unit

type export_desc =

| Export_func of indice option

| Export_table of indice option

| Export_mem of indice option

| Export_global of indice option

val pp_export_desc : Stdlib.Format.formatter -> export_desc -> unit

type export = {

name : string;

desc : export_desc;

}

val pp_export : Stdlib.Format.formatter -> export -> unit

type type_def = string option * func_type

val pp_type_def : Stdlib.Format.formatter -> (string option * ((string option * val_type) list * val_type list)) -> unit

val type_def_eq : (string option * (('a * val_type) list * val_type list)) -> (string option * (('a * val_type) list * val_type list)) -> bool

val pp_start : Stdlib.Format.formatter -> indice -> unit

type global = {

typ : global_type;

init : expr Annotated.t;

id : string option;

}

val pp_global : Stdlib.Format.formatter -> global -> Prelude.Unit.t

type data_mode =

| Data_passive

| Data_active of indice option * expr Annotated.t

val pp_data_mode : Stdlib.Format.formatter -> data_mode -> Prelude.Unit.t

type data = {

id : string option;

init : string;

mode : data_mode;

}

val pp_data : Stdlib.Format.formatter -> data -> Prelude.Unit.t

type elem_mode =

| Elem_passive

| Elem_active of indice option * expr Annotated.t

| Elem_declarative

val pp_elem_mode : Stdlib.Format.formatter -> elem_mode -> Prelude.Unit.t

type elem = {

id : string option;

typ : ref_type;

init : expr Annotated.t list;

mode : elem_mode;

}

val pp_elem_expr : Stdlib.Format.formatter -> expr Annotated.t -> Prelude.Unit.t

val pp_elem : Stdlib.Format.formatter -> elem -> Prelude.Unit.t

type module_field =

| MType of type_def

| MGlobal of global

| MTable of table

| MMem of mem

| MFunc of func

| MElem of elem

| MData of data

| MStart of indice

| MImport of import

| MExport of export

val pp_module_field : Stdlib.Format.formatter -> module_field -> Prelude.Unit.t

type modul = {

id : string option;

fields : module_field list;

}

val pp_modul : Stdlib.Format.formatter -> modul -> unit