Module Owi.Ws

Synchronized LIFO Stacks

include Work_ds_intf.S

type !'a t

val make : unit -> 'a t

Create a new queue

val push : 'a -> Prio.t -> 'a t -> unit

Add a new element to the queue

val pop : 'a t -> bool -> 'a option

Get an element from the queue. The boolean shall be true to atomically start a new pledge (cf make_pledge) while popping.

val make_pledge : 'a t -> unit

Make a new pledge, ie indicate that new elements may be pushed to the queue and that calls to pop should block waitting for them.

val end_pledge : 'a t -> unit

End one pledge.

val fail : 'a t -> unit

Mark the queue as failed: all threads trying to pop from it will get no element.

val read_as_seq : 'a t -> finalizer:(unit -> unit) -> 'a Prelude.Seq.t

Pop all elements from the queue in a lazy Seq.t,

val work_while : ('a -> ((Prio.t * 'a) -> unit) -> unit) -> 'a t -> unit