woi

owi is an OCaml toolchain to work with WebAssembly. It provides an interpreter as an executable and a library.

To get started, have a look at the quickstart section of the README.

API

Swhid.Parse

Swhid.Types

Swhid.Compile

Swhid.Check

Swhid.Simplify

Swhid.Typecheck

Swhid.Link

Swhid.Log

Swhid.Interpret

Swhid.Script

Swhid.Pp

Swhid.Value

Private API

You shouldn't have to use any of these modules, they're used internally only.

Swhid.Convert

Swhid.Float32

Swhid.Float64

Swhid.Int32

Swhid.Int64

Swhid.Lexer

Swhid.Menhir_parser

Swhid.Spectest

Swhid.Wutf8