Directory ppx_deriving/runtime

Libraries

The following libraries are found in this directory.

ppx_deriving.runtime

Modules

The following is a list of all of the modules found at this filesystem path.

Ppx_deriving_runtime A module collecting all predefined OCaml types, exceptions and modules operating on them, so that ppx_deriving plugins operate in a well-defined environment.