Package ocaml-compiler-libs

Library ocaml-compiler-libs.common

The entry point of this library is the module: Ocaml_common.

Library ocaml-compiler-libs.shadow

The entry point of this library is the module: Ocaml_shadow.