Module Owi

module Check : sig ... end

Initial check done on a module.

module Compile : sig ... end

Utility functions to compile a module until a given step.

module Convert : sig ... end

Various conversion functions between i32, i64, f32 and f64.

module Float32 : sig ... end

Custom Float32 module for Wasm.

module Float64 : sig ... end

Custom Float64 module for Wasm.

module Int32 : sig ... end

Custom Int32 module for Wasm.

module Int64 : sig ... end

Custom Int64 module for Wasm.

module Interpret : sig ... end

Module to interpret a linked module.

module Lexer : sig ... end

Module for Wasm lexing.

module Link : sig ... end

Module to link a simplified/extern module and producing a runnable module along with a link state.

module Log : sig ... end

Module to enable or disable the printing of debug logs.

module Menhir_parser : sig ... end

module Parse : sig ... end

Module providing functions to parse a wasm script from various kind of inputs.

module Pp : sig ... end

Module to pretty print the various types such as text module or simplified module.

module Script : sig ... end

Module to execute a full Wasm script.

module Simplify : sig ... end

Module to simplify a text module. It contains the types of simplified modules.

module Spectest : sig ... end

The `spectest` module, to run script from the official test suite.

module Stack : sig ... end

runtime stack module

module Typecheck : sig ... end

Module to typecheck a simplified module.

module Types : sig ... end

Types of text modules, as produced by the parser. Many parts of it are reused later in the interpreter.

module Value : sig ... end

Module to define externref values in OCaml. You should look in the `example` directory to understand how to use this before reading the code...

module Wutf8 : sig ... end

Utility functions to work with utf8.