Module Owi

module Assigned : sig ... end

module Check : sig ... end

Initial check done on a module.

module Compile : sig ... end

Utility functions to compile a module until a given step.

module Convert : sig ... end

Various conversion functions between i32, i64, f32 and f64.

module Float32 : sig ... end

Custom Float32 module for Wasm.

module Float64 : sig ... end

Custom Float64 module for Wasm.

module Global : sig ... end

module Grouped : sig ... end

module Imported : sig ... end

module Indexed : sig ... end

module Int32 : sig ... end

Custom Int32 module for Wasm.

module Int64 : sig ... end

Custom Int64 module for Wasm.

module Interpret : sig ... end

Module to interpret a linked module.

module Lexer : sig ... end

Module for Wasm lexing.

module Link : sig ... end

Module to link a simplified/extern module and producing a runnable module along with a link state.

module Link_env : sig ... end

module Log : sig ... end

Module to enable or disable the printing of debug logs.

module Memory : sig ... end

module Menhir_parser : sig ... end

module Named : sig ... end

module Optimize : sig ... end

Optimize module

module Parse : sig ... end

Module providing functions to parse a wasm script from various kind of inputs.

module Pp : sig ... end

Module to pretty print the various types such as text module or simplified module.

module Result : sig ... end

module Rewrite : sig ... end

module Runtime : sig ... end

module Script : sig ... end

Module to execute a full Wasm script.

module Simplified : sig ... end

module Simplified_types : sig ... end

module Spectest : sig ... end

The `spectest` module, to run script from the official test suite.

module Stack : sig ... end

runtime stack module

module String_map : sig ... end

module Symbolic : sig ... end

module Syntax : sig ... end

module Table : sig ... end

module Typecheck : sig ... end

Module to typecheck a simplified module.

module Types : sig ... end

module Value : sig ... end

Module to define externref values in OCaml. You should look in the `example` directory to understand how to use this before reading the code...

module Wutf8 : sig ... end

Utility functions to work with utf8.