Module Owi

module Assigned : sig ... end

module C_instrumentor : sig ... end

module C_share : sig ... end

module C_share_site : sig ... end

module Check : sig ... end

Initial check done on a module.

module Choice_intf : sig ... end

module Cmd_c : sig ... end

module Cmd_fmt : sig ... end

module Cmd_opt : sig ... end

module Cmd_run : sig ... end

module Cmd_script : sig ... end

module Cmd_sym : sig ... end

module Cmd_validate : sig ... end

module Compile : sig ... end

Utility functions to compile a module until a given step.

module Concrete : sig ... end

module Concrete_choice : sig ... end

module Concrete_global : sig ... end

runtime global

module Concrete_memory : sig ... end

module Concrete_table : sig ... end

runtime table

module Concrete_value : sig ... end

Module to define externref values in OCaml. You should look in the `example` directory to understand how to use this before reading the code...

module Convert : sig ... end

Various conversion functions between i32, i64, f32 and f64.

module Env_id : sig ... end

module Float32 : sig ... end

Custom Float32 module for Wasm.

module Float64 : sig ... end

Custom Float64 module for Wasm.

module Format : sig ... end

module Func_id : sig ... end

module Func_intf : sig ... end

module Grouped : sig ... end

module Imported : sig ... end

module Indexed : sig ... end

module Int32 : sig ... end

Custom Int32 module for Wasm.

module Int64 : sig ... end

Custom Int64 module for Wasm.

module Interpret : sig ... end

module Interpret_intf : sig ... end

module Lexer : sig ... end

Module for Wasm lexing.

module Link : sig ... end

Module to link a simplified/extern module and producing a runnable module along with a link state.

module Link_env : sig ... end

module Log : sig ... end

Module to enable or disable the printing of debug logs.

module Menhir_parser : sig ... end

module Named : sig ... end

module Optimize : sig ... end

Optimize module

module Parse : sig ... end

Module providing functions to parse a wasm script from various kind of inputs.

module Result : sig ... end

module Rewrite : sig ... end

module Runtime : sig ... end

module Script : sig ... end

Module to execute a full Wasm script.

module Simplified : sig ... end

module Simplified_types : sig ... end

module Solver : sig ... end

module Spectest : sig ... end

The `spectest` module, to run script from the official test suite.

module Stack : sig ... end

module String_map : sig ... end

module Symbolic : sig ... end

module Symbolic_choice : sig ... end

module Symbolic_global : sig ... end

module Symbolic_memory : sig ... end

module Symbolic_table : sig ... end

module Symbolic_value : sig ... end

module Syntax : sig ... end

module Text : sig ... end

module Thread : sig ... end

module Tracing : sig ... end

module Trap : sig ... end

module Typecheck : sig ... end

Module to typecheck a simplified module.

module Types : sig ... end

module V : sig ... end

module Value_intf : sig ... end

module Wutf8 : sig ... end

Utility functions to work with utf8.