Module Owi

module Annot : sig ... end

module Annotated : sig ... end

module Assigned : sig ... end

module Binary : sig ... end

module Binary_encoder : sig ... end

module Binary_parser : sig ... end

module Binary_to_text : sig ... end

module Binary_types : sig ... end

module Binary_validate : sig ... end

Module to typecheck a simplified module.

module Choice_intf : sig ... end

module Cmd_c : sig ... end

module Cmd_call_graph : sig ... end

module Cmd_cfg : sig ... end

module Cmd_conc : sig ... end

module Cmd_cpp : sig ... end

module Cmd_fmt : sig ... end

module Cmd_instrument : sig ... end

module Cmd_iso : sig ... end

module Cmd_replay : sig ... end

module Cmd_run : sig ... end

module Cmd_rust : sig ... end

module Cmd_script : sig ... end

module Cmd_sym : sig ... end

module Cmd_tinygo : sig ... end

module Cmd_utils : sig ... end

module Cmd_validate : sig ... end

module Cmd_version : sig ... end

module Cmd_wasm2wat : sig ... end

module Cmd_wat2wasm : sig ... end

module Cmd_zig : sig ... end

module Code_generator : sig ... end

module Compile : sig ... end

Utility functions to compile a module until a given step.

module Concolic : sig ... end

module Concolic_choice : sig ... end

module Concolic_extern_func : sig ... end

module Concolic_memory : sig ... end

module Concolic_value : sig ... end

module Concolic_wasm_ffi : sig ... end

module Concrete : sig ... end

module Concrete_choice : sig ... end

module Concrete_extern_func : sig ... end

module Concrete_global : sig ... end

runtime global

module Concrete_memory : sig ... end

module Concrete_table : sig ... end

runtime table

module Concrete_value : sig ... end

module Concrete_wasm_ffi : sig ... end

module Contract : sig ... end

module Convert : sig ... end

Various conversion functions between i32, i64, f32 and f64.

module Env_id : sig ... end

module Float32 : sig ... end

Custom Float32 module for Wasm.

module Float64 : sig ... end

Custom Float64 module for Wasm.

module Func_id : sig ... end

module Func_intf : sig ... end

module Graph : sig ... end

module Grouped : sig ... end

module Imported : sig ... end

module Indexed : sig ... end

module Int32 : sig ... end

Custom Int32 module for Wasm.

module Int64 : sig ... end

Custom Int64 module for Wasm.

module Interpret : sig ... end

module Interpret_intf : sig ... end

module Kind : sig ... end

module Link : sig ... end

Module to link a binary/extern module and producing a runnable module along with a link state.

module Link_env : sig ... end

module Minimalist_symbolic : sig ... end

module Minimalist_symbolic_choice : sig ... end

module Named : sig ... end

module Parse : sig ... end

Module providing functions to parse a wasm script from various kind of inputs.

module Pq_imperative : sig ... end

module Prio : sig ... end

module Priority_queue : sig ... end

module Result : sig ... end

module Rewrite : sig ... end

module Runtime : sig ... end

module Script : sig ... end

Module to execute a full Wasm script.

module Sexp : sig ... end

module Share : sig ... end

module Solver : sig ... end

module Spec : sig ... end

module Spectest : sig ... end

The `spectest` module, to run script from the official test suite.

module Stack : sig ... end

module String_map : sig ... end

module Symbol_scope : sig ... end

module Symbolic : sig ... end

module Symbolic_choice : sig ... end

module Symbolic_choice_intf : sig ... end

module Symbolic_choice_with_memory : sig ... end

module Symbolic_choice_without_memory : sig ... end

module Symbolic_functor : sig ... end

module Symbolic_global : sig ... end

module Symbolic_memory : sig ... end

module Symbolic_memory_concretizing : sig ... end

module Symbolic_memory_intf : sig ... end

module Symbolic_memory_make : sig ... end

module Symbolic_path_condition : sig ... end

module Symbolic_table : sig ... end

module Symbolic_value : sig ... end

module Symbolic_wasm_ffi : sig ... end

module Synchronizer : sig ... end

module Syntax : sig ... end

module Text : sig ... end

module Text_lexer : sig ... end

Module for Wasm lexing.

module Text_parser : sig ... end

module Text_validate : sig ... end

Initial check done on a module.

module Thread : sig ... end

module Thread_intf : sig ... end

module Thread_with_memory : sig ... end

module Thread_without_memory : sig ... end

module Types : sig ... end

module Union_find : sig ... end

module V : sig ... end

Module to define externref values in OCaml. You should look in the `example` directory to understand how to use this before reading the code...

module V128 : sig ... end

module Value_intf : sig ... end

module Wasm_ffi_intf : sig ... end

module Work_ds_intf : sig ... end

module Wpq : sig ... end

module Wq : sig ... end

Synchronized FIFO queues

module Ws : sig ... end

Synchronized LIFO Stacks

module Wutf8 : sig ... end

Utility functions to work with utf8.