Module Owi.Types

exception Trap of string

exception Parse_fail of string

type yes =

| Yes

type no =

| No

type with_string_indices = < string_indices : yes >

type without_string_indices = < string_indices : no >

type with_ind_bt = < raw_bt : yes >

type without_ind_bt = < raw_bt : no >

type text = < with_string_indices ; with_ind_bt >

type binary = < without_string_indices ; without_ind_bt >

val sp : Stdlib.Format.formatter -> unit -> unit

type _ indice =

| Text : string -> < with_string_indices.. > indice

| Raw : int -> < .. > indice

val pp_id : Stdlib.Format.formatter -> string -> unit

val pp_id_opt : Stdlib.Format.formatter -> string option -> unit

val pp_indice : Stdlib.Format.formatter -> 'kind indice -> unit

val compare_indice : 'a indice -> 'a indice -> int

val indice_eq : 'a indice -> 'a indice -> bool

val pp_indice_opt : Stdlib.Format.formatter -> 'a indice option -> unit

val pp_indices : Stdlib.Format.formatter -> 'a indice list -> unit

type nonrec num_type =

| I32

| I64

| F32

| F64

val pp_num_type : Stdlib.Format.formatter -> num_type -> unit

val num_type_eq : num_type -> num_type -> bool

val compare_num_type : num_type -> num_type -> int

type nullable =

| No_null

| Null

val pp_nullable : Stdlib.Format.formatter -> nullable -> unit

type nonrec packed_type =

| I8

| I16

val pp_packed_type : Stdlib.Format.formatter -> packed_type -> unit

val packed_type_eq : packed_type -> packed_type -> bool

type nonrec mut =

| Const

| Var

val pp_mut : Stdlib.Format.formatter -> mut -> unit

val is_mut : mut -> bool

type nonrec nn =

| S32

| S64

val pp_nn : Stdlib.Format.formatter -> nn -> unit

type nonrec sx =

| U

| S

val pp_sx : Stdlib.Format.formatter -> sx -> unit

type nonrec iunop =

| Clz

| Ctz

| Popcnt

val pp_iunop : Stdlib.Format.formatter -> iunop -> unit

type nonrec funop =

| Abs

| Neg

| Sqrt

| Ceil

| Floor

| Trunc

| Nearest

val pp_funop : Stdlib.Format.formatter -> funop -> unit

type nonrec ibinop =

| Add

| Sub

| Mul

| Div of sx

| Rem of sx

| And

| Or

| Xor

| Shl

| Shr of sx

| Rotl

| Rotr

val pp_ibinop : Stdlib.Format.formatter -> ibinop -> unit

type nonrec fbinop =

| Add

| Sub

| Mul

| Div

| Min

| Max

| Copysign

val pp_fbinop : Stdlib.Format.formatter -> fbinop -> unit

type nonrec itestop =

| Eqz

val pp_itestop : Stdlib.Format.formatter -> itestop -> unit

type nonrec irelop =

| Eq

| Ne

| Lt of sx

| Gt of sx

| Le of sx

| Ge of sx

val pp_irelop : Stdlib.Format.formatter -> irelop -> Prelude.Unit.t

type nonrec frelop =

| Eq

| Ne

| Lt

| Gt

| Le

| Ge

val frelop : Stdlib.Format.formatter -> frelop -> Prelude.Unit.t

type nonrec memarg = {

offset : Int32.t;

align : Int32.t;

}

val pp_memarg : Stdlib.Format.formatter -> memarg -> unit

type nonrec limits = {

min : int;

max : int option;

}

val pp_limits : Stdlib.Format.formatter -> limits -> unit

type nonrec mem = string option * limits

val pp_mem : Stdlib.Format.formatter -> (string option * limits) -> unit

type nonrec final =

| Final

| No_final

val pp_final : Stdlib.Format.formatter -> final -> unit

val final_eq : final -> final -> bool

Structure

Types

type 'a heap_type =

| Any_ht

| None_ht

| Eq_ht

| I31_ht

| Struct_ht

| Array_ht

| Func_ht

| No_func_ht

| Extern_ht

| No_extern_ht

| Def_ht of 'a indice

val pp_heap_type : Stdlib.Format.formatter -> 'a heap_type -> unit

val pp_heap_type_short : Stdlib.Format.formatter -> 'a heap_type -> unit

val heap_type_eq : 'a heap_type -> 'b heap_type -> bool

val compare_heap_type : 'a heap_type -> 'b heap_type -> int

type nonrec 'a ref_type = nullable * 'a heap_type

val pp_ref_type : Stdlib.Format.formatter -> (nullable * 'a heap_type) -> unit

val ref_type_eq : (nullable * 'a heap_type) -> (nullable * 'b heap_type) -> bool

val compare_ref_type : (nullable * 'a heap_type) -> (nullable * 'b heap_type) -> int

type nonrec 'a val_type =

| Num_type of num_type

| Ref_type of 'a ref_type

val pp_val_type : Stdlib.Format.formatter -> 'a val_type -> unit

val val_type_eq : 'a val_type -> 'b val_type -> bool

val compare_val_type : 'a val_type -> 'b val_type -> int

type nonrec 'a param = string option * 'a val_type

val pp_param : Stdlib.Format.formatter -> (string option * 'a val_type) -> unit

val param_eq : ('a * 'b val_type) -> ('c * 'd val_type) -> bool

val compare_param : ('a * 'b val_type) -> ('c * 'd val_type) -> int

type nonrec 'a param_type = 'a param list

val pp_param_type : Stdlib.Format.formatter -> (string option * 'a val_type) list -> unit

val param_type_eq : ('a * 'b val_type) list -> ('a * 'b val_type) list -> bool

val compare_param_type : ('a * 'b val_type) list -> ('a * 'b val_type) list -> int

type nonrec 'a result_type = 'a val_type list

val pp_result_ : Stdlib.Format.formatter -> 'a val_type -> unit

val pp_result_type : Stdlib.Format.formatter -> 'a val_type list -> unit

val result_type_eq : 'a val_type list -> 'a val_type list -> bool

val compare_result_type : 'a val_type list -> 'a val_type list -> int

val with_space_list : (Stdlib.Format.formatter -> 'a list -> unit) -> Stdlib.Format.formatter -> 'a list -> unit

type nonrec 'a func_type = 'a param_type * 'a result_type

val pp_func_type : Stdlib.Format.formatter -> ((string option * 'a val_type) list * 'b val_type list) -> unit

val func_type_eq : (('a * 'b val_type) list * 'c val_type list) -> (('a * 'b val_type) list * 'c val_type list) -> bool

type 'a block_type =

| Bt_ind : 'a indice -> < with_ind_bt.. > as 'a block_type

| Bt_raw : ('a indice option * 'a func_type) -> < .. > as 'a block_type

val pp_block_type : Stdlib.Format.formatter -> 'kind block_type -> unit

val pp_block_type_opt : Stdlib.Format.formatter -> 'a block_type option -> unit

val compare_func_type : (('a * 'b val_type) list * 'c val_type list) -> (('a * 'b val_type) list * 'c val_type list) -> int

type nonrec 'a table_type = limits * 'a ref_type

val pp_table_type : Stdlib.Format.formatter -> (limits * (nullable * 'a heap_type)) -> unit

type nonrec 'a global_type = mut * 'a val_type

val pp_global_type : Stdlib.Format.formatter -> (mut * 'a val_type) -> unit

type nonrec 'a extern_type =

| Func of string option * 'a func_type

| Table of string option * 'a table_type

| Mem of string option * limits

| Global of string option * 'a global_type

Instructions

type 'a instr =

| I32_const of Int32.t

| I64_const of Int64.t

| F32_const of Float32.t

| F64_const of Float64.t

| I_unop of nn * iunop

| F_unop of nn * funop

| I_binop of nn * ibinop

| F_binop of nn * fbinop

| I_testop of nn * itestop

| I_relop of nn * irelop

| F_relop of nn * frelop

| I_extend8_s of nn

| I_extend16_s of nn

| I64_extend32_s

| I32_wrap_i64

| I64_extend_i32 of sx

| I_trunc_f of nn * nn * sx

| I_trunc_sat_f of nn * nn * sx

| F32_demote_f64

| F64_promote_f32

| F_convert_i of nn * nn * sx

| I_reinterpret_f of nn * nn

| F_reinterpret_i of nn * nn

| Ref_null of 'a heap_type

| Ref_is_null

| Ref_i31

| Ref_func of 'a indice

| Ref_as_non_null

| Ref_cast of nullable * 'a heap_type

| Ref_test of nullable * 'a heap_type

| Ref_eq

| Drop

| Select of 'a val_type list option

| Local_get of 'a indice

| Local_set of 'a indice

| Local_tee of 'a indice

| Global_get of 'a indice

| Global_set of 'a indice

| Table_get of 'a indice

| Table_set of 'a indice

| Table_size of 'a indice

| Table_grow of 'a indice

| Table_fill of 'a indice

| Table_copy of 'a indice * 'a indice

| Table_init of 'a indice * 'a indice

| Elem_drop of 'a indice

| I_load of nn * memarg

| F_load of nn * memarg

| I_store of nn * memarg

| F_store of nn * memarg

| I_load8 of nn * sx * memarg

| I_load16 of nn * sx * memarg

| I64_load32 of sx * memarg

| I_store8 of nn * memarg

| I_store16 of nn * memarg

| I64_store32 of memarg

| Memory_size

| Memory_grow

| Memory_fill

| Memory_copy

| Memory_init of 'a indice

| Data_drop of 'a indice

| Nop

| Unreachable

| Block of string option * 'a block_type option * 'a expr

| Loop of string option * 'a block_type option * 'a expr

| If_else of string option * 'a block_type option * 'a expr * 'a expr

| Br of 'a indice

| Br_if of 'a indice

| Br_table of 'a indice array * 'a indice

| Br_on_cast of 'a indice * 'a ref_type * 'a ref_type

| Br_on_cast_fail of 'a indice * nullable * 'a heap_type

| Br_on_non_null of 'a indice

| Br_on_null of 'a indice

| Return

| Return_call of 'a indice

| Return_call_indirect of 'a indice * 'a block_type

| Return_call_ref of 'a block_type

| Call of 'a indice

| Call_indirect of 'a indice * 'a block_type

| Call_ref of 'a indice

| Array_get of 'a indice

| Array_get_u of 'a indice

| Array_len

| Array_new of 'a indice

| Array_new_data of 'a indice * 'a indice

| Array_new_default of 'a indice

| Array_new_elem of 'a indice * 'a indice

| Array_new_fixed of 'a indice * int

| Array_set of 'a indice

| I31_get_u

| I31_get_s

| Struct_get of 'a indice * 'a indice

| Struct_get_s of 'a indice * 'a indice

| Struct_new of 'a indice

| Struct_new_default of 'a indice

| Struct_set of 'a indice * 'a indice

| Extern_externalize

| Extern_internalize

and 'a expr = 'a instr list

val pp_newline : Stdlib.Format.formatter -> unit -> unit

val pp_instr : 'a instr Fmt.t

val pp_expr : Stdlib.Format.formatter -> 'a expr -> Prelude.Unit.t

val iter_expr : ('a instr -> 'b) -> 'a expr -> unit

val iter_instr : ('a instr -> 'b) -> 'a instr -> unit

type 'a func = {

type_f : 'a block_type;

locals : 'a param list;

body : 'a expr;

id : string option;

}

val pp_local : Stdlib.Format.formatter -> (string option * 'a val_type) -> unit

val pp_locals : Stdlib.Format.formatter -> (string option * 'a val_type) list -> unit

val pp_func : 'kind. Prelude.Format.formatter -> 'kind func -> unit

val pp_funcs : Stdlib.Format.formatter -> 'a func list -> unit

type 'a table = string option * 'a table_type

val pp_table : Stdlib.Format.formatter -> (string option * (limits * (nullable * 'a heap_type))) -> unit

type 'a import_desc =

| Import_func of string option * 'a block_type

| Import_table of string option * 'a table_type

| Import_mem of string option * limits

| Import_global of string option * 'a global_type

val import_desc : Stdlib.Format.formatter -> 'a import_desc -> Prelude.Unit.t

type 'a import = {

modul : string;
(*
The name of the module from which the import is done
*)

name : string;
(*
The name of the importee in its module of origin
*)

desc : 'a import_desc;
(*
If this import_desc first field is Some s, the importee is made available under name s, else it can only be used via its numerical index.
*)

}

val pp_import : Stdlib.Format.formatter -> 'a import -> Prelude.Unit.t

type 'a export_desc =

| Export_func of 'a indice option

| Export_table of 'a indice option

| Export_mem of 'a indice option

| Export_global of 'a indice option

val pp_export_desc : Stdlib.Format.formatter -> 'a export_desc -> unit

type 'a export = {

name : string;

desc : 'a export_desc;

}

val pp_export : Stdlib.Format.formatter -> text export -> unit

type 'a storage_type =

| Val_storage_t of 'a val_type

| Val_packed_t of packed_type

val pp_storage_type : Stdlib.Format.formatter -> 'a storage_type -> unit

val storage_type_eq : 'a storage_type -> 'b storage_type -> bool

type 'a field_type = mut * 'a storage_type

val pp_field_type : Stdlib.Format.formatter -> (mut * 'a storage_type) -> unit

val field_type_eq : (mut * 'a storage_type) -> (mut * 'b storage_type) -> bool

type 'a struct_field = string option * 'a field_type list

val pp_fields : Stdlib.Format.formatter -> (mut * 'a storage_type) list -> unit

val pp_struct_field : Stdlib.Format.formatter -> (string option * (mut * 'a storage_type) list) -> unit

val struct_field_eq : ('a * (mut * 'b storage_type) list) -> ('c * (mut * 'b storage_type) list) -> bool

type 'a struct_type = 'a struct_field list

val pp_struct_type : Stdlib.Format.formatter -> (string option * (mut * 'a storage_type) list) list -> unit

val struct_type_eq : ('a * (mut * 'b storage_type) list) list -> ('a * (mut * 'b storage_type) list) list -> bool

val pp_array_type : Stdlib.Format.formatter -> (mut * 'a storage_type) -> unit

type 'a str_type =

| Def_struct_t of 'a struct_type

| Def_array_t of 'a field_type

| Def_func_t of 'a func_type

val str_type : Stdlib.Format.formatter -> 'a str_type -> unit

val str_type_eq : 'a str_type -> 'a str_type -> bool

val compare_str_type : 'a str_type -> 'a str_type -> int

type 'a sub_type = final * 'a indice list * 'a str_type

val pp_sub_type : Stdlib.Format.formatter -> (final * 'a indice list * 'b str_type) -> unit

val sub_type_eq : (final * 'a indice list * 'b str_type) -> (final * 'a indice list * 'b str_type) -> bool

type 'a type_def = string option * 'a sub_type

val pp_type_def_no_indent : Stdlib.Format.formatter -> (string option * (final * 'a indice list * 'b str_type)) -> unit

val pp_type_def : Stdlib.Format.formatter -> (string option * (final * 'a indice list * 'b str_type)) -> unit

val type_def_eq : (string option * (final * 'a indice list * 'b str_type)) -> (string option * (final * 'a indice list * 'b str_type)) -> bool

type 'a rec_type = 'a type_def list

val pp_rec_type : Stdlib.Format.formatter -> (string option * (final * 'a indice list * 'b str_type)) list -> unit

val rec_type_eq : (string option * (final * 'a indice list * 'b str_type)) list -> (string option * (final * 'a indice list * 'b str_type)) list -> bool

val pp_start : Stdlib.Format.formatter -> 'a indice -> unit

type 'a const =

| Const_I32 of Int32.t

| Const_I64 of Int64.t

| Const_F32 of Float32.t

| Const_F64 of Float64.t

| Const_null of 'a heap_type

| Const_host of int

| Const_extern of int

| Const_array

| Const_eq

| Const_i31

| Const_struct

val pp_const : Stdlib.Format.formatter -> 'a const -> unit

val pp_consts : Stdlib.Format.formatter -> 'a const list -> unit