owi

owi is an OCaml toolchain to work with WebAssembly. It provides an interpreter as an executable and a library.

To get started, have a look at the quickstart section of the README.

API

Owi.Parse Module providing functions to parse a wasm script from various kind of inputs.

Owi.Types

Owi.Compile Utility functions to compile a module until a given step.

Owi.Check Initial check done on a module.

Owi.Simplify

Owi.Typecheck Module to typecheck a simplified module.

Owi.Link Module to link a simplified/extern module and producing a runnable module along with a link state.

Owi.Log Module to enable or disable the printing of debug logs.

Owi.Interpret

Owi.Script Module to execute a full Wasm script.

Owi.Pp

Owi.Value

Private API

You shouldn't have to use any of these modules, they're used internally only.

Owi.Convert Various conversion functions between i32, i64, f32 and f64.

Owi.Float32 Custom Float32 module for Wasm.

Owi.Float64 Custom Float64 module for Wasm.

Owi.Int32 Custom Int32 module for Wasm.

Owi.Int64 Custom Int64 module for Wasm.

Owi.Lexer Module for Wasm lexing.

Owi.Menhir_parser

Owi.Spectest The `spectest` module, to run script from the official test suite.

Owi.Wutf8 Utility functions to work with utf8.